pyproject.toml
[project]
name = "open-cite"
version = "0.1.0"
description = "Open Cite project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "databricks-sdk>=0.36.0",
    "databricks-sql-connector>=4.0.0",
    "mlflow>=2.17.0",
    "flask>=3.0.0",
    "gunicorn>=21.0.0",
    "boto3>=1.34.0",
    "hypercorn>=0.17.0",
    "asgiref>=3.7.0",
    "python-socketio>=5.10.0",
    "opentelemetry-proto>=1.20.0",
    "protobuf>=4.25.0",
    "pyvis>=0.3.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9"
]

[project.scripts]
opencite = "open_cite.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/open_cite"]
